Initial ACPI support
Added bare-bones ACPI support. Very few features supported, but it supports all the ACPI features that Windows Vista needs for its ACPI driver.
Also fixed 32-bit default I/O reads (again). The size of the result was being truncated to 16 bits.
Added some more configuration options, including planned CD-ROM support

+++ src/io.c
@@ -32,7 +32,7 @@ uint32_t io_default_readw(uint32_t port)
 }
 uint32_t io_default_readd(uint32_t port)
 {
-    uint16_t res = io_readb(port);
+    uint32_t res = io_readb(port);
     res |= io_readb(port + 1) << 8;
     res |= io_readb(port + 2) << 16;
     return res | io_readb(port + 3) << 24;
